done :: amounts set to u64

+++ file.rs
@@ -5,7 +5,7 @@ use anchor_lang::prelude::*;
 pub struct Lending {
   pub lender: Pubkey,
   pub borrower: Pubkey,
-  pub amount: u8, //@note :: dummy flaw that will be fixed by deeydos via PR to u64
+  pub amount: u64, //@note :: dummy flaw that will be fixed by deeydos via PR to u64
   pub interest_bps: u16,
   pub borrowed_date: i64,
   pub grace_period: i64,
